handle transcripts in videos
